+++ crates/ln-dlc-node/src/node/dlc_channel.rs
@@ -21,7 +21,6 @@ use dlc_manager::contract::ContractDescriptor;
 use dlc_manager::DlcChannelId;
 use dlc_manager::Oracle;
 use dlc_manager::Storage;
-use dlc_messages::channel::Reject;
 use dlc_messages::ChannelMessage;
 use dlc_messages::Message;
 use time::OffsetDateTime;
@@ -85,28 +84,6 @@ impl<S: TenTenOneStorage + 'static, N: LnDlcStorage + Sync + Send + 'static> Nod
             }
         })
         .await?
-    }
-
-    pub fn reject_dlc_channel_offer(&self, channel_id: &DlcChannelId) -> Result<()> {
-        let channel_id_hex = hex::encode(channel_id);
-
-        let channel = self
-            .dlc_manager
-            .get_store()
-            .get_channel(channel_id)?
-            .with_context(|| format!("Couldn't find channel by id. {}", channel_id.to_hex()))?;
-
-        tracing::info!(channel_id = %channel_id_hex, "Rejecting DLC channel offer");
-
-        self.event_handler.publish(NodeEvent::SendDlcMessage {
-            peer: channel.get_counter_party_id(),
-            msg: Message::Channel(ChannelMessage::Reject(Reject {
-                channel_id: channel.get_id(),
-                timestamp: OffsetDateTime::now_utc().unix_timestamp() as u64,
-            })),
-        })?;
-
-        Ok(())
     }
 
     pub fn accept_dlc_channel_offer(&self, channel_id: &DlcChannelId) -> Result<()> {