+++ crates/tests-e2e/tests/open_position.rs
@@ -23,7 +23,7 @@ fn dummy_order() -> NewOrder {
 #[tokio::test(flavor = "multi_thread")]
 #[ignore = "need to be run with 'just e2e' command"]
 async fn can_open_position() {
-    let test = TestSetup::new_after_funding().await;
+    let test = TestSetup::new_after_funding(None).await;
     let app = &test.app;
 
     let order = dummy_order();