+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Feat(webapp): Show order history
+- Fix: Add reject dlc channel, settle and renew offer
 
 ## [1.8.4] - 2024-01-31